+++ Data/Scripts/Indev/Block placing.lua
@@ -1,0 +1,7 @@
+function Behavior:Awake()
+    
+end
+
+function Behavior:Update()
+    
+end