+++ Data/Scripts/Menu/Mouse.lua
@@ -1,0 +1,6 @@
+function Behavior:Awake()
+end
+
+function Behavior:Update()
+    CS.Input.UnlockMouse()
+end